move prepareStructureSecondaryRollMacro into macros.ts
